# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AfriConnect API - Super-app for West Africa
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ── WALLET ────────────────────────────────────────────────────────
...
  /courses:
    get:
      operationId: listCourses
      tags: [education]
      summary: List available courses
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: level
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
  /courses/{id}:
    get:
      operationId: getCourse
      tags: [education]
      summary: Get course details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDetail"
  /courses/{id}/enroll:
    post:
      operationId: enrollCourse
      tags: [education]
      summary: Enroll in a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollInput"
      responses:
        "201":
          description: Enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /admin/users/{id}/status:
    patch:
      operationId: updateUserStatus
      tags: [admin]
      summary: Update user status (ban/activate)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserStatusUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
components:
  schemas:
...
    User:
      type: object
      required: [id, firstName, lastName, email, phone, country, role, status, createdAt]
      properties:
        id:         { type: integer }
        firstName:  { type: string }
        lastName:   { type: string }
        email:      { type: string }
        phone:      { type: string }
        country:    { type: string }
        avatarUrl:  { type: ["string", "null"] }
        coverUrl:   { type: ["string", "null"] }
        bio:        { type: ["string", "null"] }
        role:       { type: string, enum: [user, admin] }
        status:     { type: string, enum: [active, banned] }
        createdAt:  { type: string }
    UserUpdate:
      type: object
      properties:
        firstName: { type: string }
        lastName:  { type: string }
        phone:     { type: string }
        country:   { type: string }
        bio:       { type: string }
        avatarUrl: { type: string }
        coverUrl:  { type: string }
    UserStatusUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [active, banned] }

    # ── WALLET ────────────────────────────────────────────────────
...
    Course:
      type: object
      required: [id, title, category, level, instructorId, duration, enrollmentsCount, createdAt]
      properties:
        id:               { type: integer }
        title:            { type: string }
        description:      { type: ["string", "null"] }
        category:         { type: string }
        level:            { type: string, enum: [beginner, intermediate, advanced] }
        imageUrl:         { type: ["string", "null"] }
        instructorId:     { type: integer }
        duration:         { type: integer }
        enrollmentsCount: { type: integer }
        isFree:           { type: boolean }
        price:            { type: ["number", "null"] }
        currency:         { type: ["string", "null"] }
        createdAt:        { type: string }
    CourseDetail:
      type: object
      required: [id, title, category, level, instructorId, duration, enrollmentsCount, lessons, createdAt]
      properties:
        id:               { type: integer }
        title:            { type: string }
        description:      { type: ["string", "null"] }
        category:         { type: string }
        level:            { type: string }
        imageUrl:         { type: ["string", "null"] }
        instructorId:     { type: integer }
        duration:         { type: integer }
        enrollmentsCount: { type: integer }
        isFree:           { type: boolean }
        price:            { type: ["number", "null"] }
        currency:         { type: ["string", "null"] }
        lessons:
          type: array
          items:
            $ref: "#/components/schemas/Lesson"
        createdAt: { type: string }
    Lesson:
      type: object
      required: [id, courseId, title, duration, order]
      properties:
        id:       { type: integer }
        courseId: { type: integer }
        title:    { type: string }
        content:  { type: ["string", "null"] }
        videoUrl: { type: ["string", "null"] }
        duration: { type: integer }
        order:    { type: integer }
    Enrollment:
      type: object
      required: [id, courseId, userId, progress, status, enrolledAt]
      properties:
        id:         { type: integer }
        courseId:   { type: integer }
        userId:     { type: integer }
        progress:   { type: integer }
        status:     { type: string, enum: [active, completed] }
        enrolledAt: { type: string }
    EnrollInput:
      type: object
      required: [userId]
      properties:
        userId: { type: integer }

    # ── SOCIAL ────────────────────────────────────────────────────
...