# lib/api-zod/src/generated/types/course.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AfriConnect API - Super-app for West Africa
 * OpenAPI spec version: 0.1.0
 */
import type { CourseLevel } from './courseLevel';

export interface Course {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  category: string;
  level: CourseLevel;
  /** @nullable */
  imageUrl?: string | null;
  instructorId: number;
  duration: number;
  enrollmentsCount: number;
  isFree?: boolean;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  currency?: string | null;
  createdAt: string;
}