# lib/api-zod/src/generated/types/enrollment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AfriConnect API - Super-app for West Africa
 * OpenAPI spec version: 0.1.0
 */
import type { EnrollmentStatus } from './enrollmentStatus';

export interface Enrollment {
  id: number;
  courseId: number;
  userId: number;
  progress: number;
  status: EnrollmentStatus;
  enrolledAt: string;
}