# lib/api-zod/src/generated/types/lesson.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AfriConnect API - Super-app for West Africa
 * OpenAPI spec version: 0.1.0
 */

export interface Lesson {
  id: number;
  courseId: number;
  title: string;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  videoUrl?: string | null;
  duration: number;
  order: number;
}